@@ -11,33 +11,25 @@
     "clear": "docusaurus clear",
     "serve": "docusaurus serve",
     "write-translations": "docusaurus write-translations",
-    "write-heading-ids": "docusaurus write-heading-ids",
-    "typecheck": "tsc"
+    "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.1",
-    "@docusaurus/plugin-client-redirects": "^2.4.1",
-    "@docusaurus/preset-classic": "^2.4.1",
-    "@docusaurus/theme-classic": "^2.4.1",
-    "@mdx-js/react": "^1.6.22",
-    "@phosphor-icons/react": "^2.0.6",
-    "@uiw/react-markdown-preview": "4.2.2",
-    "autoprefixer": "^10.4.16",
-    "axios": "^1.6.2",
-    "clsx": "^1.2.1",
-    "docusaurus-plugin-sass": "^0.2.2",
-    "docusaurus-prince-pdf": "^1.1.1",
-    "postcss": "^8.4.31",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "sass": "^1.55.0",
-    "tailwindcss": "^3.3.5"
+    "@docusaurus/core": "3.0.1",
+    "@docusaurus/plugin-client-redirects": "^3.0.1",
+    "@docusaurus/preset-classic": "3.0.1",
+    "@ionic-internal/ionic-ds": "^8.0.0",
+    "@mdx-js/react": "^3.0.0",
+    "@phosphor-icons/react": "^2.0.15",
+    "clsx": "^2.0.0",
+    "docusaurus-plugin-sass": "^0.2.5",
+    "prism-react-renderer": "^2.3.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "sass": "^1.69.5"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.1",
-    "@tsconfig/docusaurus": "^1.0.5",
-    "typescript": "^4.7.4"
+    "@docusaurus/module-type-aliases": "3.0.1",
+    "@docusaurus/types": "3.0.1"
   },
   "browserslist": {
     "production": [
@@ -46,12 +38,12 @@
       "not op_mini all"
     ],
     "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
+      "last 3 chrome version",
+      "last 3 firefox version",
+      "last 5 safari version"
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
