@@ -320,4 +320,4 @@ const sidebars = {
   ],
 };
 
-module.exports = sidebars;
+export default sidebars;
