@@ -11,7 +11,7 @@ const Card = ({ title, icon, link, children }) => (
         {icon && <div className="icon-wrapper">{icon}</div>}
         <div className="content">
           <h3 className="title">{title}</h3>
-          <p>{children}</p>
+          {children}
         </div>
         {/* <div className="right">
           <ArrowRight size={24} color="#0070F3" weight="regular" />
