@@ -1,6 +1,6 @@
 import ExecutionEnvironment from "@docusaurus/ExecutionEnvironment";
 import React, { useState } from "react";
-import "./style.scss";
+import styles from "./styles.module.scss";
 
 const Feedback = ({ label }) => {
   if (!ExecutionEnvironment.canUseDOM) {
@@ -19,23 +19,23 @@ const Feedback = ({ label }) => {
   };
 
   return (
-    <div className="feedback">
-      <h5 className="title">Was this page useful?</h5>
+    <div className={styles.Feedback}>
+      <h5 className={styles.Title}>Was this page useful?</h5>
 
       {haveVoted ? (
         <p>Thanks for letting us know!</p>
       ) : (
         <>
-          <span title="Happy" data-doc-useful="yes" className="action" onClick={() => giveFeedback("yes")}>
-            <span className="iconx">
+          <span title="Happy" data-doc-useful="yes" className={styles.Action} onClick={() => giveFeedback("yes")}>
+            <span className={styles.Iconx}>
               <svg viewBox="0 0 20 20" focusable="false">
                 <path d="M12.36 12.23C11.6915 12.769 10.8587 13.0629 10 13.0629C9.14131 13.0629 8.30849 12.769 7.64 12.23C7.43579 12.0603 7.17251 11.9786 6.90808 12.003C6.64365 12.0274 6.39974 12.1558 6.23 12.36C6.06027 12.5642 5.9786 12.8275 6.00298 13.0919C6.02736 13.3563 6.15579 13.6003 6.36 13.77C7.38134 14.6226 8.66957 15.0896 10 15.0896C11.3304 15.0896 12.6187 14.6226 13.64 13.77C13.8442 13.6003 13.9726 13.3563 13.997 13.0919C14.0214 12.8275 13.9397 12.5642 13.77 12.36C13.686 12.2589 13.5828 12.1753 13.4665 12.1141C13.3501 12.0528 13.2229 12.0151 13.0919 12.003C12.8275 11.9786 12.5642 12.0603 12.36 12.23ZM7 9C7.19779 9 7.39113 8.94135 7.55557 8.83147C7.72002 8.72159 7.8482 8.56541 7.92388 8.38268C7.99957 8.19996 8.01937 7.99889 7.98079 7.80491C7.9422 7.61093 7.84696 7.43275 7.70711 7.29289C7.56726 7.15304 7.38908 7.0578 7.19509 7.01921C7.00111 6.98063 6.80005 7.00043 6.61732 7.07612C6.43459 7.15181 6.27842 7.27998 6.16853 7.44443C6.05865 7.60888 6 7.80222 6 8C6 8.26522 6.10536 8.51957 6.2929 8.70711C6.48043 8.89464 6.73479 9 7 9ZM13 7C12.8022 7 12.6089 7.05865 12.4444 7.16853C12.28 7.27841 12.1518 7.43459 12.0761 7.61732C12.0004 7.80004 11.9806 8.00111 12.0192 8.19509C12.0578 8.38907 12.153 8.56725 12.2929 8.70711C12.4328 8.84696 12.6109 8.9422 12.8049 8.98079C12.9989 9.01937 13.2 8.99957 13.3827 8.92388C13.5654 8.84819 13.7216 8.72002 13.8315 8.55557C13.9414 8.39112 14 8.19778 14 8C14 7.73478 13.8946 7.48043 13.7071 7.29289C13.5196 7.10536 13.2652 7 13 7ZM10 0C8.02219 0 6.08879 0.58649 4.4443 1.6853C2.79981 2.78412 1.51809 4.3459 0.761209 6.17317C0.00433284 8.00043 -0.193701 10.0111 0.192152 11.9509C0.578004 13.8907 1.53041 15.6725 2.92894 17.0711C4.32746 18.4696 6.10929 19.422 8.0491 19.8079C9.98891 20.1937 11.9996 19.9957 13.8268 19.2388C15.6541 18.4819 17.2159 17.2002 18.3147 15.5557C19.4135 13.9112 20 11.9778 20 10C20 8.68678 19.7413 7.38642 19.2388 6.17317C18.7363 4.95991 17.9997 3.85752 17.0711 2.92893C16.1425 2.00035 15.0401 1.26375 13.8268 0.761205C12.6136 0.258658 11.3132 0 10 0ZM10 18C8.41775 18 6.87104 17.5308 5.55544 16.6518C4.23985 15.7727 3.21447 14.5233 2.60897 13.0615C2.00347 11.5997 1.84504 9.99113 2.15372 8.43928C2.4624 6.88743 3.22433 5.46197 4.34315 4.34315C5.46197 3.22433 6.88743 2.4624 8.43928 2.15372C9.99113 1.84504 11.5997 2.00346 13.0615 2.60896C14.5233 3.21447 15.7727 4.23984 16.6518 5.55544C17.5308 6.87103 18 8.41775 18 10C18 12.1217 17.1572 14.1566 15.6569 15.6569C14.1566 17.1571 12.1217 18 10 18Z"></path>
               </svg>
             </span>
             Yes
           </span>
-          <span title="Disappointed" data-doc-useful="no" className="action" onClick={() => giveFeedback("no")}>
-            <span className="iconx">
+          <span title="Disappointed" data-doc-useful="no" className={styles.Action} onClick={() => giveFeedback("no")}>
+            <span className={styles.Iconx}>
               <svg viewBox="0 0 20 20" focusable="false">
                 <path d="M6.36 13.33C6.15844 13.4993 6.03176 13.7412 6.00742 14.0033C5.98308 14.2654 6.06306 14.5265 6.23 14.73C6.31395 14.8313 6.41705 14.915 6.5334 14.9763C6.64975 15.0377 6.77706 15.0755 6.90803 15.0875C7.03901 15.0996 7.17108 15.0857 7.29669 15.0467C7.42229 15.0076 7.53896 14.9442 7.64 14.86C8.30849 14.321 9.14131 14.0271 10 14.0271C10.8587 14.0271 11.6915 14.321 12.36 14.86C12.4611 14.9442 12.5777 15.0076 12.7033 15.0467C12.8289 15.0857 12.961 15.0996 13.092 15.0875C13.223 15.0755 13.3503 15.0377 13.4666 14.9763C13.583 14.915 13.6861 14.8313 13.77 14.73C13.9369 14.5265 14.0169 14.2654 13.9926 14.0033C13.9683 13.7412 13.8416 13.4993 13.64 13.33C12.6081 12.5006 11.3239 12.0484 10 12.0484C8.67609 12.0484 7.39188 12.5006 6.36 13.33ZM7 9C7.19779 9 7.39113 8.94135 7.55557 8.83147C7.72002 8.72159 7.8482 8.56541 7.92388 8.38268C7.99957 8.19996 8.01937 7.99889 7.98079 7.80491C7.9422 7.61093 7.84696 7.43275 7.70711 7.29289C7.56726 7.15304 7.38908 7.0578 7.19509 7.01921C7.00111 6.98063 6.80005 7.00043 6.61732 7.07612C6.43459 7.15181 6.27842 7.27998 6.16853 7.44443C6.05865 7.60888 6 7.80222 6 8C6 8.26522 6.10536 8.51957 6.2929 8.70711C6.48043 8.89464 6.73479 9 7 9ZM10 0C8.02219 0 6.08879 0.58649 4.4443 1.6853C2.79981 2.78412 1.51809 4.3459 0.761209 6.17317C0.00433284 8.00043 -0.193701 10.0111 0.192152 11.9509C0.578004 13.8907 1.53041 15.6725 2.92894 17.0711C4.32746 18.4696 6.10929 19.422 8.0491 19.8079C9.98891 20.1937 11.9996 19.9957 13.8268 19.2388C15.6541 18.4819 17.2159 17.2002 18.3147 15.5557C19.4135 13.9112 20 11.9778 20 10C20 8.68678 19.7413 7.38642 19.2388 6.17317C18.7363 4.95991 17.9997 3.85752 17.0711 2.92893C16.1425 2.00035 15.0401 1.26375 13.8268 0.761205C12.6136 0.258658 11.3132 0 10 0ZM10 18C8.41775 18 6.87104 17.5308 5.55544 16.6518C4.23985 15.7727 3.21447 14.5233 2.60897 13.0615C2.00347 11.5997 1.84504 9.99113 2.15372 8.43928C2.4624 6.88743 3.22433 5.46197 4.34315 4.34315C5.46197 3.22433 6.88743 2.4624 8.43928 2.15372C9.99113 1.84504 11.5997 2.00346 13.0615 2.60896C14.5233 3.21447 15.7727 4.23984 16.6518 5.55544C17.5308 6.87103 18 8.41775 18 10C18 12.1217 17.1572 14.1566 15.6569 15.6569C14.1566 17.1571 12.1217 18 10 18ZM13 7C12.8022 7 12.6089 7.05865 12.4444 7.16853C12.28 7.27841 12.1518 7.43459 12.0761 7.61732C12.0004 7.80004 11.9806 8.00111 12.0192 8.19509C12.0578 8.38907 12.153 8.56725 12.2929 8.70711C12.4328 8.84696 12.6109 8.9422 12.8049 8.98079C12.9989 9.01937 13.2 8.99957 13.3827 8.92388C13.5654 8.84819 13.7216 8.72002 13.8315 8.55557C13.9414 8.39112 14 8.19778 14 8C14 7.73478 13.8946 7.48043 13.7071 7.29289C13.5196 7.10536 13.2652 7 13 7Z"></path>
               </svg>
